requirements.txt
Flask==2.3.3
pandas==2.1.1
qrcode==7.4.2
scikit-learn==1.3.0
Pillow==10.0.1
pymongo
numpy
google-generativeai
python-dotenv>=1.0.1